requirements.txt
alembic==1.3.1
asn1crypto==1.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
configparser==4.0.2
cryptography==2.8
databricks-cli==0.9.1
docker==4.2.0
entrypoints==0.3
Flask==1.1.1
gitdb2==2.0.6
GitPython==3.0.5
gorilla==0.3.0
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
kafka-python==2.0.0
Mako==1.1.1
MarkupSafe==1.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mlflow==1.6.0
numpy==1.18.1
pandas==1.0.1
prometheus-client==0.7.1
prometheus-flask-exporter==0.12.2
protobuf==3.11.3
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
PyYAML==5.3
querystring-parser==1.2.4
requests==2.22.0
simplejson==3.17.0
six==1.14.0
smmap2==2.0.5
SQLAlchemy==1.3.13
sqlparse==0.3.0
tabulate==0.8.6
urllib3==1.25.8
websocket-client==0.57.0
Werkzeug==1.0.0
